requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
django-mysql==4.12.0
djangorestframework==3.14.0
flake8==6.1.0
isort==5.12.0
kombu==5.3.4
mccabe==0.7.0
mysqlclient==2.2.0
prompt-toolkit==3.0.41
pycodestyle==2.11.1
pyflakes==3.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
pytest==7.4.3
pytest-django==4.7.0